added E2E tests
